Add files via upload
Microsoft Office and in particular Microsoft Access have available Libraries, and in those modules and classes.  This is my attempt to write a detailed book about those modules and classes, and within classes how to use events, both sinking events in a class as well as raising events from classes.